Housekeeping: Embracing modularity and all that
Usual cleanup commit, minor refactoring on the ErrorList component.
