+++ src/components/cc-form/components/index.ts
@@ -1,0 +1,2 @@
+export { Error } from './Error'
+export { Input } from './Input'
